requirements.txt
bs4
requests
dateparser
pandas
pyyaml
flask
flask-API
click
flask-compress
flask-cors
